+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased]
 
+### New features
+
+- `thingAsMarkdown()`: a function that takes a Thing and returns a string representation of it
+  that can assist in debugging issues.
+
 ## [0.2.0] - 2020-08-27
 
 ### New features